Refactor LOADFIX out of dos_programs
